+++ src/main/java/org/terasology/pathfinding/model/PathCache.java
@@ -1,22 +1,9 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
 package org.terasology.pathfinding.model;
 
 import com.google.common.collect.Maps;
-import org.terasology.navgraph.WalkableBlock;
+import org.terasology.pathfinding.navgraph.WalkableBlock;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -25,7 +12,7 @@ import java.util.Map;
  * @author synopia
  */
 public class PathCache {
-    private Map<WalkableBlock, Map<WalkableBlock, Path>> paths = Maps.newHashMap();
+    private final Map<WalkableBlock, Map<WalkableBlock, Path>> paths = Maps.newHashMap();
 
     public Path getCachedPath(WalkableBlock from, WalkableBlock to) {
         Map<WalkableBlock, Path> fromMap = paths.get(from);
