+++ src/main/java/org/terasology/pathfinding/model/Pathfinder.java
@@ -1,23 +1,10 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
 package org.terasology.pathfinding.model;
 
 import com.google.common.collect.Lists;
-import org.terasology.navgraph.NavGraphSystem;
-import org.terasology.navgraph.WalkableBlock;
+import org.terasology.pathfinding.navgraph.NavGraphSystem;
+import org.terasology.pathfinding.navgraph.WalkableBlock;
 
 import java.util.Collections;
 import java.util.List;
@@ -26,9 +13,9 @@ import java.util.List;
  *
  */
 public class Pathfinder {
-    private HAStar haStar;
-    private PathCache cache;
-    private NavGraphSystem world;
+    private final HAStar haStar;
+    private final PathCache cache;
+    private final NavGraphSystem world;
 
     public Pathfinder(NavGraphSystem world, LineOfSight lineOfSight) {
         this.world = world;
