+++ src/main/java/org/terasology/pathfinding/navgraph/FloorFinder.java
@@ -1,19 +1,6 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.navgraph;
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.pathfinding.navgraph;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -29,11 +16,11 @@ import java.util.Set;
  * @author synopia
  */
 public class FloorFinder {
-    private List<Region> regions = Lists.newArrayList();
-    private Map<Region, Integer> count = Maps.newHashMap();
-    private List<Sweep> sweeps = Lists.newArrayList();
-    private Map<WalkableBlock, Region> regionMap = Maps.newHashMap();
-    private Map<WalkableBlock, Sweep> sweepMap = Maps.newHashMap();
+    private final List<Region> regions = Lists.newArrayList();
+    private final Map<Region, Integer> count = Maps.newHashMap();
+    private final List<Sweep> sweeps = Lists.newArrayList();
+    private final Map<WalkableBlock, Region> regionMap = Maps.newHashMap();
+    private final Map<WalkableBlock, Sweep> sweepMap = Maps.newHashMap();
 
 
     public FloorFinder() {
