+++ src/main/java/org/terasology/pathfinding/navgraph/Sweep.java
@@ -1,0 +1,12 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.pathfinding.navgraph;
+
+/**
+ * @author synopia
+ */
+public class Sweep {
+    public Region region;
+    public Region neighbor;
+    public int neighborCount;
+}
