+++ src/main/java/org/terasology/pathfinding/navgraph/WalkableBlock.java
@@ -1,19 +1,6 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.terasology.navgraph;
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.pathfinding.navgraph;
 
 import org.terasology.math.geom.Vector3i;
 
@@ -23,7 +10,7 @@ import org.terasology.math.geom.Vector3i;
 public class WalkableBlock {
     public WalkableBlock[] neighbors = new WalkableBlock[8];
     public Floor floor;
-    private Vector3i position;
+    private final Vector3i position;
 
     public WalkableBlock(int x, int z, int height) {
         position = new Vector3i(x, height, z);
