+++ src/test/java/org/terasology/pathfinding/AStarTest.java
@@ -1,25 +1,12 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
 package org.terasology.pathfinding;
 
 import com.google.common.collect.Lists;
 import org.junit.Assert;
 import org.junit.Test;
-import org.terasology.navgraph.BitMap;
 import org.terasology.pathfinding.model.AStar;
+import org.terasology.pathfinding.navgraph.BitMap;
 
 import java.util.Collections;
 import java.util.List;
